+++ src/module/enhanced-conditions/enhanced-condition-macro.js
@@ -1,27 +1,36 @@
+const { HandlebarsApplicationMixin, ApplicationV2 } = foundry.applications.api;
+
 /**
  * Enhanced Condition Macro Config Application
  */
-export default class EnhancedConditionMacroConfig extends FormApplication {
-	constructor(object, options) {
-		super(object, options);
-
-		this.object = this.object ?? {};
+export default class EnhancedConditionMacroConfig extends HandlebarsApplicationMixin(ApplicationV2) {
+	constructor(object, options = {}) {
+		super(options);
+		this.object = object ?? {};
 		this.object.macros = this.object.macros ?? [];
-
 		this.initialObject = foundry.utils.duplicate(this.object);
 	}
 
-	static get defaultOptions() {
-		return foundry.utils.mergeObject(super.defaultOptions, {
-			id: "cub-enhanced-condition-macro-config",
-			title: game.i18n.localize("CLT.ENHANCED_CONDITIONS.MacroConfig.Title"),
-			template: "modules/ironsworn-impacts/templates/enhanced-condition-macro-config.hbs",
-			classes: ["sheet"],
+	static DEFAULT_OPTIONS = {
+		id: "cub-enhanced-condition-macro-config",
+		window: { title: "CLT.ENHANCED_CONDITIONS.MacroConfig.Title" },
+		position: { width: 400, height: "auto" },
+		classes: ["sheet"],
+		form: {
+			handler: EnhancedConditionMacroConfig.#onSubmit,
+			submitOnChange: false,
 			closeOnSubmit: false
-		});
-	}
+		},
+		tag: "form"
+	};
 
-	async getData() {
+	static PARTS = {
+		form: {
+			template: "modules/ironsworn-impacts/templates/enhanced-condition-macro-config.hbs"
+		}
+	};
+
+	async _prepareContext(_options) {
 		const conditionMacros = this.object.macros;
 		const applyMacroId = conditionMacros.find((m) => m.type === "apply")?.id;
 		const removeMacroId = conditionMacros.find((m) => m.type === "remove")?.id;
@@ -43,30 +52,26 @@ export default class EnhancedConditionMacroConfig extends FormApplication {
 		}
 		compendiumMacros.sort((a, b) => a.name.localeCompare(b.name));
 
-		const macroChoices = [...worldMacros, ...compendiumMacros];
-
 		return {
 			condition: this.object,
 			applyMacroId,
 			removeMacroId,
-			macroChoices
+			macroChoices: [...worldMacros, ...compendiumMacros]
 		};
 	}
 
-	async _updateObject(event, formData) {
+	static async #onSubmit(_event, _form, formData) {
+		const data = formData.object;
 		this.object.macros = [];
 
-		for (const field in formData) {
-			const type = field.split("-").slice(-1)
-				.pop() ?? "";
-			const tempMacro = { id: formData[field], type: type };
-			this.object.macros.push(tempMacro);
+		for (const field in data) {
+			const type = field.split("-").slice(-1).pop() ?? "";
+			this.object.macros.push({ id: data[field], type });
 		}
 
 		const map = game.clt.conditions;
 		const newMap = foundry.utils.duplicate(map);
-
-		let conditionIndex = newMap.findIndex((c) => c.id === this.object.id);
+		const conditionIndex = newMap.findIndex((c) => c.id === this.object.id);
 		newMap[conditionIndex] = this.object;
 		await game.settings.set("ironsworn-impacts", "activeConditionMap", newMap);
 		this.close();