+++ CHANGELOG.md
@@ -4,4 +4,4 @@
 
 ## [1.0.00](https://github.com/jendave/ironsworn-impacts/blob/main/CHANGELOG.md) (2026-04-29)
 
-* Initial release.
+* Initial release. The module was created to provide system-specific support for Ironsworn and Starforged. Also, v13 support was updated and v14 support was added.